refactorisation de la classe auth.php
