+++ auth/nyxei/classes/NotificationMessages.php
@@ -15,7 +15,8 @@ class NotificationMessages
     private static $message = [
         'alert_message_login_failed' => "Alert: Multiple Failed Login Attempts",
         'failed_login_attempts_message' => "User %s has had %d failed login attempts",
-        'ldap_connection_failed' => "Failed to connect to LDAP. Verify the host and LDAPS support."
+        'ldap_connection_failed' => "Failed to connect to LDAP. Verify the host and LDAPS support.",
+        'invalid_credentials' => "Invalid Credentials",
     ];
 
     public static function getMessage(string $key,array $params = []): string
